+++ crates/uv/tests/it/help.rs
@@ -541,9 +541,7 @@ fn help_subsubcommand() {
               Distributions can be read from a local directory by using the `file://` URL scheme.
 
           --python-downloads-json-url <PYTHON_DOWNLOADS_JSON_URL>
-              URL pointing to JSON of custom Python installations.
-              
-              Note that currently, only local paths are supported.
+              URL pointing to JSON of custom Python installations
 
       -r, --reinstall
               Reinstall the requested Python version, if it's already installed.